+++ gtk2_ardour/editor.cc
@@ -31,7 +31,12 @@
 
 /*
  * Modified by Apollo AI Lab, Inc. on 2026-03-16
- * Summary of changes: Added WebView2 chat panel integration for LabDAWg AI assistant
+ * Summary of changes: Wired ChatWebView widget into the main editor window;
+ * added show/hide toggle and layout integration.
+ *
+ * Modified by Apollo AI Lab, Inc. on 2026-06-06
+ * Summary of changes: Removed ChatWebView widget references following the
+ * removal of the in-process chat panel.
  */
 
 /* Note: public Editor methods are documented in public_editor.h */
@@ -59,7 +64,6 @@
 #include "pbd/memento_command.h"
 #include "pbd/unknown_type.h"
 #include "pbd/unwind.h"
-#include "pbd/timersub.h"
 
 #include <glibmm/datetime.h> /*for playlist group_id */
 #include <glibmm/miscutils.h>
@@ -119,9 +123,6 @@
 #include "editing.h"
 #include "editing_convert.h"
 #include "editor.h"
-#ifdef PLATFORM_WINDOWS
-#include "chat_webview.h"
-#endif
 #include "editor_cursors.h"
 #include "editor_drag.h"
 #include "editor_group_tabs.h"
@@ -570,6 +571,8 @@ Editor::Editor ()
 
 	_notebook_tab1.set_name ("tab button");
 	_notebook_tab2.set_name ("tab button");
+	_notebook_tab1.set_text_color (0xffffffff);
+	_notebook_tab2.set_text_color (0xffffffff);
 
 	/* Pick up some settings we need to cache, early */
 
@@ -578,33 +581,72 @@ Editor::Editor ()
 	editor_summary_pane.set_check_divider_position (true);
 	editor_summary_pane.add (edit_packer);
 
-	Button* summary_arrow_left = manage (new Button);
-	summary_arrow_left->add (*manage (new Arrow (ARROW_LEFT, SHADOW_NONE)));
-	summary_arrow_left->signal_pressed().connect (sigc::hide_return (sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), LEFT)));
-	summary_arrow_left->signal_released().connect (sigc::mem_fun (*this, &Editor::scroll_release));
+	ArdourButton* summary_arrow_left = manage (new ArdourButton);
+	summary_arrow_left->set_corner_mask(ArdourButton::LEFT);
+	summary_arrow_left->set_icon(ArdourIcon::ArrowLeft);
+	summary_arrow_left->set_act_on_release(false);
+	summary_arrow_left->signal_button_press_event().connect (sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), LEFT), false);
+	summary_arrow_left->signal_button_release_event().connect (sigc::mem_fun (*this, &Editor::scroll_release), false);
 
-	Button* summary_arrow_right = manage (new Button);
-	summary_arrow_right->add (*manage (new Arrow (ARROW_RIGHT, SHADOW_NONE)));
-	summary_arrow_right->signal_pressed().connect (sigc::hide_return (sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), RIGHT)));
-	summary_arrow_right->signal_released().connect (sigc::mem_fun (*this, &Editor::scroll_release));
+	ArdourButton* summary_arrow_right = manage (new ArdourButton);
+	summary_arrow_right->set_corner_mask(ArdourButton::RIGHT);
+	summary_arrow_right->set_icon(ArdourIcon::ArrowRight);
+	summary_arrow_right->set_act_on_release(false);
+	summary_arrow_right->signal_button_press_event().connect (sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), RIGHT), false);
+	summary_arrow_right->signal_button_release_event().connect (sigc::mem_fun (*this, &Editor::scroll_release), false);
 
-	VBox* summary_arrows_left = manage (new VBox);
-	summary_arrows_left->pack_start (*summary_arrow_left);
+	Gtk::EventBox* summary_left_spacer = manage (new Gtk::EventBox); // extra h-space before the summary
+	summary_left_spacer->set_size_request(4, -1);
+	summary_left_spacer->show();
+	Gtk::EventBox* summary_bottom_spacer = manage (new Gtk::EventBox); // extra v-space after the summary
+	summary_bottom_spacer->set_size_request(-1, 3);
+	summary_bottom_spacer->show();
 
-	VBox* summary_arrows_right = manage (new VBox);
-	summary_arrows_right->pack_start (*summary_arrow_right);
+	VBox* summary_zoom_vbox = manage (new Gtk::VBox);
+	HBox* summary_zoom_hbox = manage (new Gtk::HBox);
 
-	Gtk::Frame* summary_frame = manage (new Gtk::Frame);
-	summary_frame->set_shadow_type (Gtk::SHADOW_ETCHED_IN);
+	HBox* summary_toolbar = manage (new Gtk::HBox);
+	summary_toolbar->pack_start(*summary_arrow_left, true, true, 0);
+	summary_toolbar->pack_start(zoom_out_button, true, true, 0);
+	summary_toolbar->pack_start(zoom_in_button, true, true, 0);
+	summary_toolbar->pack_start(full_zoom_button, true, true, 0);
+	summary_toolbar->pack_start(*summary_arrow_right, true, true, 0);
 
-	summary_frame->add (*_summary);
-	summary_frame->show ();
+	Glib::RefPtr<SizeGroup> grp = SizeGroup::create (Gtk::SIZE_GROUP_BOTH);
+	grp->add_widget(*summary_arrow_right);
+	grp->add_widget(full_zoom_button);
+	grp->add_widget(zoom_in_button);
+	grp->add_widget(zoom_out_button);
+	grp->add_widget(*summary_arrow_left);
 
-	_summary_hbox.pack_start (*summary_arrows_left, false, false);
+	summary_zoom_hbox->pack_end(*summary_toolbar, false, false, 0);
+	summary_zoom_vbox->pack_end(*summary_zoom_hbox, false, false, 0);
+
+	summary_arrow_left->set_corner_mask(ArdourButton::NONE);
+	zoom_out_button.set_corner_mask(ArdourButton::NONE);
+	zoom_in_button.set_corner_mask(ArdourButton::NONE);
+	full_zoom_button.set_corner_mask(ArdourButton::NONE);
+	summary_arrow_right->set_corner_mask(ArdourButton::NONE);
+
+	summary_arrow_left->set_border_mask(ArdourButton::HIDE_BOTTOM);
+	zoom_out_button.set_border_mask(ArdourButton::HIDE_LEFT | ArdourButton::HIDE_RIGHT | ArdourButton::HIDE_BOTTOM);
+	zoom_in_button.set_border_mask(ArdourButton::HIDE_BOTTOM);
+	full_zoom_button.set_border_mask(ArdourButton::HIDE_LEFT | ArdourButton::HIDE_RIGHT | ArdourButton::HIDE_BOTTOM);
+	summary_arrow_right->set_border_mask(ArdourButton::HIDE_RIGHT | ArdourButton::HIDE_BOTTOM);
+
+	_summary->add(*summary_zoom_vbox);
+
+	Gtk::Frame* summary_frame = manage (new Gtk::Frame); // summary border
+	summary_frame->set_name("SummaryFrame");
+	summary_frame->add(*_summary);
+
+	_summary_hbox.pack_start (*summary_left_spacer, false, false);
 	_summary_hbox.pack_start (*summary_frame, true, true);
-	_summary_hbox.pack_start (*summary_arrows_right, false, false);
 
-	editor_summary_pane.add (_summary_hbox);
+	_summary_vbox.pack_start (_summary_hbox, true, true);
+	_summary_vbox.pack_start (*summary_bottom_spacer, false, false);
+
+	editor_summary_pane.add (_summary_vbox);
 
 	HBox* tabbox = manage (new HBox (true));
 	tabbox->set_spacing (3);
@@ -619,12 +661,13 @@ Editor::Editor ()
 	content_right_pane.set_drag_cursor (*_cursors->expand_left_right);
 	editor_summary_pane.set_drag_cursor (*_cursors->expand_up_down);
 
-	float fract;
-	if (!settings || !settings->get_property ("edit-vertical-pane-pos", fract) || fract > 1.0) {
-		/* initial allocation is 90% to canvas, 10% to summary */
-		fract = 0.90;
+	float size;
+	if (!settings || !settings->get_property ("edit-vertical-pane-size", size) || size < 1.0) {
+		/* initial allocation so that summary has minimum size*/
+		size = -1;
+		editor_summary_pane.set_divider (0, 0.999);
 	}
-	editor_summary_pane.set_divider (0, fract);
+	editor_summary_pane.set_absolute_divider (0, ArdourWidgets::Pane::DividerMode::AbsoluteAfter, size);
 
 	global_vpacker.set_spacing (0);
 	global_vpacker.set_border_width (0);
@@ -641,15 +684,9 @@ Editor::Editor ()
 	/* pack all the main pieces into appropriate containers from _tabbable
 	 */
 	content_app_bar.add (_application_bar);
-#ifdef PLATFORM_WINDOWS
-	_chat_webview = new ChatWebView ();
-	_right_pane.add (_editor_list_vbox);
-	_right_pane.add (*_chat_webview);
-	_right_pane.set_divider (0, 0.65f);
-	content_att_right.add (_right_pane);
-#else
+
 	content_att_right.add (_editor_list_vbox);
-#endif
+
 	content_att_bottom.add (_bottom_hbox);
 	content_main_top.add (global_vpacker);
 	content_main.add (editor_summary_pane);
@@ -661,11 +698,6 @@ Editor::Editor ()
 	global_vpacker.show();
 	_bottom_hbox.show();
 	_editor_list_vbox.show_all ();
-#ifdef PLATFORM_WINDOWS
-	_chat_webview->navigate ("http://localhost:9871/");
-	_chat_webview->show_all ();
-	_right_pane.show ();
-#endif
 
 	/* register actions now so that set_state() can find them and set toggles/checks etc */
 
@@ -920,11 +952,19 @@ Editor::set_entered_regionview (RegionView* rv)
 		*/
 		sensitize_all_region_actions (true);
 	}
+
+	if (rv) {
+		set_entered_track (&rv->get_time_axis_view());
+	}
 }
 
 void
 Editor::set_entered_track (TimeAxisView* tav)
 {
+	if (entered_track == tav) {
+		return;
+	}
+
 	if (entered_track) {
 		entered_track->exited ();
 	}
@@ -2352,7 +2392,7 @@ Editor::get_state () const
 
 	node->add_child_nocopy (Tabbable::get_state());
 
-	node->set_property("edit-vertical-pane-pos", editor_summary_pane.get_divider());
+	node->set_property("edit-vertical-pane-size", editor_summary_pane.get_absolute_divider());
 
 	maybe_add_mixer_strip_width (*node);
 
@@ -2681,21 +2721,13 @@ Editor::setup_toolbar ()
 	mouse_mode_size_group->add_widget (mouse_draw_button);
 	mouse_mode_size_group->add_widget (mouse_content_button);
 
-	if (!Profile->get_mixbus()) {
-		mouse_mode_size_group->add_widget (zoom_in_button);
-		mouse_mode_size_group->add_widget (zoom_out_button);
-		mouse_mode_size_group->add_widget (full_zoom_button);
-		mouse_mode_size_group->add_widget (zoom_focus_selector);
-		mouse_mode_size_group->add_widget (tav_shrink_button);
-		mouse_mode_size_group->add_widget (tav_expand_button);
-		mouse_mode_size_group->add_widget (follow_playhead_button);
-		mouse_mode_size_group->add_widget (follow_edits_button);
-		mouse_mode_size_group->add_widget (_notebook_tab1);
-		mouse_mode_size_group->add_widget (_notebook_tab2);
-	} else {
-		mouse_mode_size_group->add_widget (zoom_preset_selector);
-		mouse_mode_size_group->add_widget (visible_tracks_selector);
-	}
+	mouse_mode_size_group->add_widget (zoom_focus_selector);
+	mouse_mode_size_group->add_widget (tav_shrink_button);
+	mouse_mode_size_group->add_widget (tav_expand_button);
+	mouse_mode_size_group->add_widget (follow_playhead_button);
+	mouse_mode_size_group->add_widget (follow_edits_button);
+	mouse_mode_size_group->add_widget (_notebook_tab1);
+	mouse_mode_size_group->add_widget (_notebook_tab2);
 
 	mouse_mode_size_group->add_widget (stretch_marker_cb);
 
@@ -2757,14 +2789,7 @@ Editor::setup_toolbar ()
 	act = ActionManager::get_action (X_("Editor"), X_("zoom-to-session"));
 	full_zoom_button.set_related_action (act);
 
-	if (ARDOUR::Profile->get_mixbus()) {
-		_zoom_box.pack_start (zoom_preset_selector, false, false);
-	} else {
-		_zoom_box.pack_start (zoom_out_button, false, false);
-		_zoom_box.pack_start (zoom_in_button, false, false);
-		_zoom_box.pack_start (full_zoom_button, false, false);
-		_zoom_box.pack_start (zoom_focus_selector, false, false);
-	}
+	_zoom_box.pack_start (zoom_focus_selector, false, false);
 
 	/* Track zoom buttons */
 	_track_box.set_spacing (2);
@@ -5195,7 +5220,7 @@ Editor::check_step_edit ()
 }
 
 bool
-Editor::scroll_press (Direction dir)
+Editor::scroll_press (GdkEventButton* ev, Direction dir)
 {
 	++_scroll_callbacks;
 
@@ -5226,7 +5251,7 @@ Editor::scroll_press (Direction dir)
 	if (!_scroll_connection.connected ()) {
 
 		_scroll_connection = Glib::signal_timeout().connect (
-			sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), dir), 100
+			sigc::bind (sigc::mem_fun (*this, &Editor::scroll_press), ev, dir), 100
 			);
 
 		_scroll_callbacks = 0;
@@ -5235,10 +5260,11 @@ Editor::scroll_press (Direction dir)
 	return true;
 }
 
-void
-Editor::scroll_release ()
+bool
+Editor::scroll_release (GdkEventButton* ev)
 {
 	_scroll_connection.disconnect ();
+	return true;
 }
 
 void
@@ -5594,8 +5620,6 @@ Editor::ui_parameter_changed (string parameter)
 	} else if (parameter == "use-note-bars-for-velocity") {
 		ArdourCanvas::Note::set_show_velocity_bars (UIConfiguration::instance().get_use_note_bars_for_velocity());
 		_track_canvas->request_redraw (_track_canvas->visible_area());
-	} else if (parameter == "use-note-color-for-velocity") {
-		/* handled individually by each MidiRegionView */
 	} else if (parameter == "show-selection-marker") {
 		update_ruler_visibility ();
 	}
@@ -5844,3 +5868,26 @@ Editor::upper_left() const
 	return get_trackview_group ()->canvas_origin ();
 }
 
+void
+Editor::toggle_main ()
+{
+	/* this code requires removing the alignment that holds
+	 * content_att_bottom in order to function
+	 */
+#if 0
+	Gtk::Box* parent = dynamic_cast<Gtk::Box*> (content_att_bottom.get_parent());
+
+	if (content_main.is_visible()) {
+		content_main_vbox.remove (content_main);
+		if (parent) {
+			parent->set_child_packing (content_att_bottom, true, true, 0);
+		}
+	} else {
+		content_main_vbox.pack_start (content_main, true, true);
+		content_main_vbox.reorder_child (content_main, 1);
+		if (parent) {
+			parent->set_child_packing (content_att_bottom, false, false, 0);
+		}
+	}
+#endif
+}